+++ src/pas/plugins/authomatic/browser/view.py
@@ -103,10 +103,8 @@ class AuthomaticView(BrowserView):
         elif provider not in cfg:
             return _("Provider not supported")
         if not self.is_anon:
-            if self.provider in self._provider_names:
-                logger.warn(
-                    "Provider %s is already connected to current "
-                    "user.", self.provider)
+            if provider in self._provider_names:
+                logger.warn(f"Provider {provider} is already connected to current user")
                 return self._redirect()
             # TODO: some sort of CSRF check might be needed, so that
             #       not an account got connected by CSRF. Research needed.
@@ -137,7 +135,7 @@ class AuthomaticView(BrowserView):
         self.request.response.redirect(
             self.context.absolute_url() + next_url
         )
-        return _("redirecting")
+        return _(f"redirecting")
 
     @property
     def is_anon(self):